deepks parallelization : gamma point, energy and force
