+++ source/Makefile.Objects
@@ -164,6 +164,8 @@ LCAO_nnr.o \
 LCAO_diago.o\
 LCAO_evolve.o\
 LCAO_descriptor.o\
+LCAO_descriptor_old.o\
+LCAO_descriptor_new.o\
 ylm.o\
 FORCE_STRESS.o\
 FORCE_gamma.o\
